LayoutTests/http/tests/loading/status-code-error-without-response-body.html

<html>
<body>
Test for <a href="https://bugs.webkit.org/show_bug.cgi?id=60875">https://bugs.webkit.org/show_bug.cgi?id=60875</a> <br>
This test verifies that FrameLoader properly handles 4xx and 5xx errors without response bodies.<br>
If this test succeeds, you should see a series of "PASS" messages and then "DONE".<br><br>

<pre id="result">
</pre>

<script>
function log(message)
{
    document.getElementById("result").innerHTML += message;
}

function done()
{
    log("\n\nDONE\n");
    if (window.layoutTestController)
        layoutTestController.notifyDone();
}

function runTest(url, expectSuccess)
{
    req = new XMLHttpRequest();
    req.onload = function() {
        log((expectSuccess ? "PASS " : "FAIL:" + req.status + " "));
        nextTest();
    }
    req.onerror = function() {
        log((expectSuccess ? "FAIL " : "PASS "));
        nextTest();
    }

    try {
        req.open("GET", url, true);
        req.send(null);
    } catch(e) {
        log("Exception: " + e.message + "\n");
    }    
}

var statusCodes = [400, 401, 402, 403, 404, 405, 406, 407, 408, 409, 410, 411, 412, 413, 414, 415, 416, 417, 500, 501, 502, 503, 504, 505]
var tests = [
    ["resources/status-code-error-with-response-body.php?statusCode=", true],
    ["resources/status-code-error-without-response-body.php?statusCode=", false]
]
var currentStatusCode = 0;
var currentTest = 0;

function nextTest()
{
    if (currentStatusCode == 0) {
        if (currentTest == 0)
            log("** Test for normal; receiving response body\n");
        else if (currentTest == 1)
            log("\n\n** Test for abnormal; not receiving response body\n");
    }

    if (currentTest >= tests.length) {
        done();
        return;
    }

    if (currentStatusCode < statusCodes.length) {
        runTest(tests[currentTest][0] + statusCodes[currentStatusCode++], tests[currentTest][1]);
        return;
    }

    currentStatusCode = 0;
    currentTest++;
    nextTest();
}

if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();

    nextTest();
} else {
    log("This test needs to be run in DumpRenderTree.\n");
}

</script>

</body>
</html>
