LayoutTests/http/tests/loading/status-code-error-without-response-body-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
Test for https://bugs.webkit.org/show_bug.cgi?id=60875 
This test verifies that FrameLoader properly handles 4xx and 5xx errors without response bodies.
If this test succeeds, you should see a series of "PASS" messages and then "DONE".

** Test for normal; receiving response body
PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS 

** Test for abnormal; not receiving response body
PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS PASS 

DONE
